+++ EQM-GQE.SHARED-RESOURCES/Interfaces/IQuestionnaireLogic.cs
@@ -9,7 +9,7 @@ namespace EQM_GQE.SHARED_RESOURCES.Interfaces
 {
     public interface IQuestionnaireLogic
     {
-        Task<long> Add(Questionnaire Questionnaire);
+        Task<int> Add(Questionnaire Questionnaire);
 
         Task<bool> Update(Questionnaire Questionnaire);
